requirements.txt
streamlit>=1.39.0
pandas>=2.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0
plotly>=5.24.0
python-dateutil>=2.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
unidecode>=1.3.0
pydeck>=0.8.0

# ============================================
# Dependencias para Piloto IA
# ============================================

# RAG y LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.5
joblib>=1.3.0
imbalanced-learn>=0.11.0

# Reportes
reportlab>=4.0.0
python-docx>=1.1.0

# Utilidades
pydantic>=2.0.0
tqdm>=4.65.0
scipy>=1.11.0
